requirements.txt
python-telegram-bot==21.*
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
apscheduler==3.10.*
pydantic-settings==2.*
openai==1.*
